+++ HISTORY.md
@@ -22,6 +22,7 @@
     * new "play frequency" commands in the Sounds category
     * pixel access primitives for bitmap and vector (!) graphics
     * new "stretch" primitive for costumes, also for flipping
+    * new "get graphic effect" reporter
     * added "neg" selector to monadic function reporter in "Operators" category
     * added "log2" selector to monadic function reporter in "Operators" category
     * added "^" reporter (power of) in the Operators category
@@ -71,6 +72,7 @@
 * Threads: new feature: new costume from list of pixels
 * Objects, Threads: added "current" to costume input slot dropdown
 * Blocks: deprecated graphic effects: "duplicate", "comic" and "confetti" 
+* Objects: added reporter for graphic effects
 
 ### 2019-04-08
 * Blocks, Objects, Threads: new "getSoundAttribute" reporter primitive